fix rssi format

+++ src/main/java/ru/r2cloud/lora/loraat/LoraAtSerialClient.java
@@ -137,7 +137,7 @@ public class LoraAtSerialClient implements LoraAtClient {
 				}
 				LoraFrame curFrame = new LoraFrame();
 				curFrame.setData(Util.hexStringToByteArray(parts[0]));
-				curFrame.setRssi(Short.parseShort(parts[1]));
+				curFrame.setRssi((short) Float.parseFloat(parts[1]));
 				curFrame.setSnr(Float.parseFloat(parts[2]));
 				curFrame.setFrequencyError(Long.parseLong(parts[3]));
 				curFrame.setTimestamp(Long.parseLong(parts[4]));
